app/requirements.txt
dspy-ai
google-cloud-aiplatform
fastapi
uvicorn[standard]
pydantic
python-dotenv
wikipedia
wikipedia-api